moving to new comp layout
