+++ comp/api/commonendpoints/commonendpointsimpl/common_endpoints.go
@@ -1,0 +1,34 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+// Package commonendpointsimpl provides common Agent API endpoints implementation
+package commonendpointsimpl
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/DataDog/datadog-agent/cmd/agent/common/signals"
+	"github.com/DataDog/datadog-agent/pkg/util/hostname"
+	"github.com/DataDog/datadog-agent/pkg/util/log"
+)
+
+func GetHostname(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	hname, err := hostname.Get(r.Context())
+	if err != nil {
+		log.Warnf("Error getting hostname: %s\n", err) // or something like this
+		hname = ""
+	}
+	j, _ := json.Marshal(hname)
+	w.Write(j)
+}
+
+func StopAgent(w http.ResponseWriter, _ *http.Request) {
+	signals.Stopper <- true
+	w.Header().Set("Content-Type", "application/json")
+	j, _ := json.Marshal("")
+	w.Write(j)
+}
